+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## v6.0 (current major)
 
+### v6.3.0
+* Added zoom segment to in-player UI and popup. 
+* Fixed keyboard zoom
+* Added additional zoom options. If you wonder how zoom options differ from crop, mess around and find out.
+* Subdomains now inherit same settings as their parent domain by default
+* Extension detects embedded content (but not always)
+* Added `www.youtube-nocookie.com` to "officially supported" list
+
 ### v6.2.5
 
 * Popup appearance changed — UI advertisement panel was moved to the popup header